Relax version constraints for dependencies
We need to use >= instead of ~> for version constraints because Chef
Policies use "dotted decimal identifier" versions that result in
integers that will never match the "normal" cookbook versions we use.

For example, the version constraint:

~> 3.2

is what we used. However, the dotted decimal version may be

30625036585770760.42672982950081083.162968991950205

The dotted decimal version is used in the "backwards compatibility
mode" for Chef Zero as it doesn't have the native Policy document API
(yet). This may result in a 412 precondition failed exception when
using the pessmistic version constraint.